django_backend/requirements.txt
django>=4.2,<5.0  
django-allauth
django-cors-headers
djangorestframework
python-dotenv
gunicorn
whitenoise